ReadyRoom
Added a ready room class. When players join a the server, they are added to a ready room. If the currently filling ready room is full, a new ready room is created.  
Players are removed from ready rooms when they disconnect.  
Messages are only sent between players within the same ready room.