Player class. Rename player.
Added a player class. Player class is created when a socket connects. Contains the socket, an id number, and a name.  
Player class will primarily be responsible for I/O to and between players, and for residence in games/Ready Rooms. Need to decide whether Ready Room should have access to player sockets.  

Added list of players to front end, and ability to rename players.